Branch from websecurity

+++ src/main/java/com/example/cateam4spring/repo/StudentRepository.java
@@ -1,10 +1,13 @@
 package com.example.cateam4spring.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
 import com.example.cateam4spring.model.Student;
+
+import javax.transaction.Transactional;
 
 public interface StudentRepository extends JpaRepository<Student, Integer> {
 	
@@ -15,7 +18,7 @@ public interface StudentRepository extends JpaRepository<Student, Integer> {
 	public Student findStudentById(Integer id);
 	
 	@Query("SELECT s.GPA FROM Student s WHERE s.id = :id")
-	public Integer findGpaById(@Param("id") Integer id);
+	public Double findGpaById(@Param("id") Integer id);
 
 	@Modifying
 	@Transactional
